+++ libraries/bio/src/helm/types.ts
@@ -2,39 +2,11 @@ import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 import * as grok from 'datagrok-api/grok';
 
-export interface IWebEditorSizes {
-  get rightwidth(): number;
-  get topheight(): number;
-  get bottomheight(): number;
-}
-
-export interface IWebEditorHelm {
-  setSequence(seq: string, format: string,
-    sugar?: string, linker?: string, append?: boolean, separator?: string): boolean;
-}
-
-export interface IWebEditorCanvas {
-  get helm(): IWebEditorHelm;
-  getHelm(ret: boolean): string;
-
-  resize(width: number, height: number): void;
-}
-
-export interface IWebEditorApp {
-  get toolbarheight(): number;
-  get canvas(): IWebEditorCanvas;
-  get structureview(): any;
-  get mex(): any;
-
-  calculateSizes(): IWebEditorSizes;
-}
-
-export interface IHelmWebEditorEditor {
-  setHelm(helm: string): void;
-}
+import '../types/jsdraw2';
+import * as JSDraw2 from 'JSDraw2';
 
 export interface IHelmWebEditor {
-  get editor(): IHelmWebEditorEditor;
+  get editor(): JSDraw2.Editor;
   get host(): HTMLDivElement;
 
   resizeEditor(width: number, height: number): void;