+++ packages/Helm/src/helm-monomer-placer.ts
@@ -2,11 +2,12 @@ import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 
+import * as org from 'org';
+import * as JSDraw2 from 'JSDraw2';
 import wu from 'wu';
 
 import {IMonomerLib, Monomer} from '@datagrok-libraries/bio/src/types/index';
 import {CellRendererBackBase} from '@datagrok-libraries/bio/src/utils/cell-renderer-back-base';
-import {IEditorMol} from '@datagrok-libraries/bio/src/types/helm-web-editor';
 
 import {getParts, parseHelm} from './utils';
 
@@ -22,7 +23,7 @@ export class HelmMonomerPlacer extends CellRendererBackBase<string> {
 
   private _allPartsList: (string[] | null)[];
   private _lengthsList: (number[] | null)[];
-  private _editorMolList: (IEditorMol | null)[];
+  private _editorMolList: (JSDraw2.IEditorMol | null)[];
 
   public monomerCharWidth: number = 7;
   public leftPadding: number = 5;
@@ -40,7 +41,7 @@ export class HelmMonomerPlacer extends CellRendererBackBase<string> {
   protected override reset(): void {
     this._allPartsList = new Array<string[] | null>(this.tableCol.length).fill(null);
     this._lengthsList = new Array<number[] | null>(this.tableCol.length).fill(null);
-    this._editorMolList = new Array<IEditorMol | null>(this.tableCol.length).fill(null);
+    this._editorMolList = new Array<JSDraw2.IEditorMol | null>(this.tableCol.length).fill(null);
   }
 
   /** Skips cell for the fallback rendering */
@@ -99,11 +100,11 @@ export class HelmMonomerPlacer extends CellRendererBackBase<string> {
     this.invalidateGrid();
   }
 
-  public setEditorMol(tableRowIndex: number, editorMol: IEditorMol) {
+  public setEditorMol(tableRowIndex: number, editorMol: JSDraw2.IEditorMol) {
     this._editorMolList![tableRowIndex] = editorMol.clone(false);
   }
 
-  public getEditorMol(tableRowIndex: number): IEditorMol | null {
+  public getEditorMol(tableRowIndex: number): JSDraw2.IEditorMol | null {
     return this._editorMolList![tableRowIndex];
   }
 }