+++ packages/Helm/webpack.config.js
@@ -36,6 +36,10 @@ module.exports = {
     'cash-dom': '$',
     'dayjs': 'dayjs',
     'wu': 'wu',
+    'scil': 'scil',
+    'org': 'org',
+    'dojo': 'dojo',
+    'JSDraw2': 'JSDraw2',
   },
   output: {
     filename: '[name].js',