+++ packages/Helm/CHANGELOG.md
@@ -5,6 +5,11 @@
 Fix Helm cell renderer to clear editor on empty cell value
 Add interactivity on monomers into HelmGridCellRenderer
 Add sync render cell from cache (Helm editor specific)
+Prioritize render queue by task consumer id and callback
+Add HelmService test
+Add Helm cell renderer test for scatter plot
+Use types of Pistoia Helm, get rid of ts-ignore
+Add handling missing monomers with Helm cell renderer
 
 ## 2.1.34 (2024-04-22)